规范架构
