+++ UiLib/src/main/res/values/strings.xml
