+++ UiLib/src/test/java/com/gpufast/uilib/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.gpufast.uilibrary;
+package com.gpufast.uilib;
 
 import org.junit.Test;
