+++ build.gradle
@@ -24,3 +24,13 @@ allprojects {
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+ext{
+    compileSdkVersion = 28
+    buildToolsVersion = "28.0.3"
+    minSdkVersion = 19
+    targetSdkVersion = 28
+    supportVersion = "28.0.0"
+    versionCode = 1
+    versionName = "1.0.0"
+}
